Bring some sidebar improvements across from comments branch
