+++ app/components/Sidebar/Right.tsx
@@ -1,27 +1,19 @@
 import { m } from "framer-motion";
 import { observer } from "mobx-react";
-import { BackIcon } from "outline-icons";
 import * as React from "react";
-import { useTranslation } from "react-i18next";
 import styled, { useTheme } from "styled-components";
 import breakpoint from "styled-components-breakpoint";
-import Button from "~/components/Button";
 import Flex from "~/components/Flex";
-import Scrollable from "~/components/Scrollable";
 import ResizeBorder from "~/components/Sidebar/components/ResizeBorder";
-import Tooltip from "~/components/Tooltip";
 import usePersistedState from "~/hooks/usePersistedState";
 
 type Props = React.HTMLAttributes<HTMLDivElement> & {
-  title: React.ReactNode;
   children: React.ReactNode;
-  onClose: React.MouseEventHandler;
   border?: boolean;
 };
 
-function RightSidebar({ title, onClose, children, border, className }: Props) {
+function Right({ children, border, className }: Props) {
   const theme = useTheme();
-  const { t } = useTranslation();
   const [width, setWidth] = usePersistedState(
     "rightSidebarWidth",
     theme.sidebarWidth
@@ -101,18 +93,7 @@ function RightSidebar({ title, onClose, children, border, className }: Props) {
       className={className}
     >
       <Position style={style} column>
-        <Header>
-          <Title>{title}</Title>
-          <Tooltip tooltip={t("Close")} shortcut="Esc" delay={500}>
-            <Button
-              icon={<ForwardIcon />}
-              onClick={onClose}
-              borderOnHover
-              neutral
-            />
-          </Tooltip>
-        </Header>
-        <Scrollable topShadow>{children}</Scrollable>
+        {children}
         <ResizeBorder
           onMouseDown={handleMouseDown}
           onDoubleClick={handleReset}
@@ -122,11 +103,6 @@ function RightSidebar({ title, onClose, children, border, className }: Props) {
     </Sidebar>
   );
 }
-
-const ForwardIcon = styled(BackIcon)`
-  transform: rotate(180deg);
-  flex-shrink: 0;
-`;
 
 const Position = styled(Flex)`
   position: fixed;
@@ -149,26 +125,4 @@ const Sidebar = styled(m.div)<{ $border?: boolean }>`
   `};
 `;
 
-const Title = styled(Flex)`
-  font-size: 16px;
-  font-weight: 600;
-  text-align: center;
-  align-items: center;
-  justify-content: flex-start;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  user-select: none;
-  overflow: hidden;
-  width: 0;
-  flex-grow: 1;
-`;
-
-const Header = styled(Flex)`
-  align-items: center;
-  position: relative;
-  padding: 16px 12px 16px 16px;
-  color: ${(props) => props.theme.text};
-  flex-shrink: 0;
-`;
-
-export default observer(RightSidebar);
+export default observer(Right);
